:recycle: File division.
