+++ src/routes/home-routes.ts
@@ -1,0 +1,9 @@
+import express from 'express';
+
+const router = express.Router();
+
+router.get('/', (req, res) => {
+  res.send('Hello wold!');
+});
+
+export default router;
