+++ src/routes/index.ts
@@ -1,0 +1,3 @@
+export { default as homeRoutes } from './home-routes';
+export { default as outgoingRoutes } from './outgoing-routes';
+export { default as twimlRoutes } from './twiml-routes';
